Change atom size to 64 bytes, from 4096 bytes
There are a few major implications. The first is that you can no longer
have sectors with only 1 atom, beacuse sectors need to have QuorumSize
hashes. Under the new system, sectors can have 0 atoms (indicating no
data), otherwise they need a minimum of QuorumSize/2 atoms, because the
first QuorumSize/2 atoms will be used to store the hash values of the
sector.

This additionally means that the storage proofs and verifications are
going to work differently.