fix network benchmark
for some reason, NewRPCServer was reporting an "address already in use"
error when running the SendMessage benchmark. Various attempts were
made to close the RPC server before opening a new one, but all failed.
Instead, the b.N variable is simply added to the port number used.
I expected this to cause dozens or hundreds of ports to be used during
the benchmark, but I was unable to catch it ever using more than one.
Not quite sure what's going on here; maybe a result of parallelism in
the benchmarking suite.