Separate wallet keys from sibling keys
Wallet keys are now completely independent from sibling keys during
bootstraping, something that enables the client to create a generic
wallet that doesn't expose the private key of the sibling. It's
important to keep them separate for anonymity reasons.

This commit is partially hacked together, and a restructuring of the
bootstrapping procedure is probably needed. Luckily, this should mostly
entail changing variable names and moving functions to different files,
with just a bit of code restructuring thrown in. Nothing as nasty as
we've done in the past.